--- misc/services/mobile-webui/mobile-webui-frontend/src/apps/pos/actions/common.js
@@ -1,4 +1,0 @@
-import { getApplicationState } from '../../index';
-import { APPLICATION_ID } from '../constants';
-
-export const getPOSApplicationState = (globalState) => getApplicationState(globalState, APPLICATION_ID);
