+++ misc/services/mobile-webui/mobile-webui-frontend/src/apps/pos/reducers/uiUtils.js
@@ -1,0 +1,6 @@
+import { determineApplicationState } from './commonUtils';
+
+export const getModalFromState = ({ applicationState, globalState }) => {
+  const applicationStateEff = determineApplicationState({ applicationState, globalState });
+  return applicationStateEff.modal;
+};
